+++ src/gen-cmd/command_list.h
@@ -57,6 +57,8 @@ SOFTWARE.
 #include "cmd_signal.h"
 #include "cmd_sleep.h"
 #include "cmd_spawn.h"
+#include "cmd_su_exec.h"
+#include "cmd_su_spawn.h"
 #include "cmd_subcmd.h"
 #include "cmd_touch_trunc.h"
 #include "cmd_wait_pid.h"
@@ -97,6 +99,8 @@ SOFTWARE.
             GLOBAL_VARDEF__SIGNAL \
             GLOBAL_VARDEF__SLEEP \
             GLOBAL_VARDEF__SPAWN \
+            GLOBAL_VARDEF__SU_EXEC \
+            GLOBAL_VARDEF__SU_SPAWN \
             GLOBAL_VARDEF__SUBCMD \
             GLOBAL_VARDEF__TOUCH_TRUNC \
             GLOBAL_VARDEF__WAIT_PID \
@@ -135,6 +139,8 @@ SOFTWARE.
             INITIALIZE__SIGNAL \
             INITIALIZE__SLEEP \
             INITIALIZE__SPAWN \
+            INITIALIZE__SU_EXEC \
+            INITIALIZE__SU_SPAWN \
             INITIALIZE__SUBCMD \
             INITIALIZE__TOUCH_TRUNC \
             INITIALIZE__WAIT_PID \
@@ -173,6 +179,8 @@ SOFTWARE.
             STARTUP_CASE__SIGNAL \
             STARTUP_CASE__SLEEP \
             STARTUP_CASE__SPAWN \
+            STARTUP_CASE__SU_EXEC \
+            STARTUP_CASE__SU_SPAWN \
             STARTUP_CASE__SUBCMD \
             STARTUP_CASE__TOUCH_TRUNC \
             STARTUP_CASE__WAIT_PID \
@@ -211,6 +219,8 @@ SOFTWARE.
             RUN_CASE__SIGNAL \
             RUN_CASE__SLEEP \
             RUN_CASE__SPAWN \
+            RUN_CASE__SU_EXEC \
+            RUN_CASE__SU_SPAWN \
             RUN_CASE__SUBCMD \
             RUN_CASE__TOUCH_TRUNC \
             RUN_CASE__WAIT_PID \
@@ -249,6 +259,8 @@ SOFTWARE.
             REQUIRES_ADDL_ARG__SIGNAL \
             REQUIRES_ADDL_ARG__SLEEP \
             REQUIRES_ADDL_ARG__SPAWN \
+            REQUIRES_ADDL_ARG__SU_EXEC \
+            REQUIRES_ADDL_ARG__SU_SPAWN \
             REQUIRES_ADDL_ARG__SUBCMD \
             REQUIRES_ADDL_ARG__TOUCH_TRUNC \
             REQUIRES_ADDL_ARG__WAIT_PID \
@@ -291,6 +303,8 @@ enum CommandIndex {
             ENUM_LIST__SIGNAL
             ENUM_LIST__SLEEP
             ENUM_LIST__SPAWN
+            ENUM_LIST__SU_EXEC
+            ENUM_LIST__SU_SPAWN
             ENUM_LIST__SUBCMD
             ENUM_LIST__TOUCH_TRUNC
             ENUM_LIST__WAIT_PID
@@ -332,6 +346,8 @@ enum CommandIndex {
             VIRTUAL_ENUM_LIST__SIGNAL
             VIRTUAL_ENUM_LIST__SLEEP
             VIRTUAL_ENUM_LIST__SPAWN
+            VIRTUAL_ENUM_LIST__SU_EXEC
+            VIRTUAL_ENUM_LIST__SU_SPAWN
             VIRTUAL_ENUM_LIST__SUBCMD
             VIRTUAL_ENUM_LIST__TOUCH_TRUNC
             VIRTUAL_ENUM_LIST__WAIT_PID